rsvp/src/utilities/requirements.txt
faker
python-ulid